backend/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
pytz==2025.2
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
django-storages==1.14.4
google-cloud-storage==2.18.0
requests==2.32.3
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
djangorestframework-simplejwt==5.3.1
